+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 * index: Do not deserialize originations matched by the `source` filter.
 * index: Wrap storage deserialization exceptions with `InvalidDataError`.
 
+### Security
+
+* hasura: Forbid using Hasura instances running vulnerable versions.
+
 ## [6.3.0] - 2022-11-15
 
 ### Added